fixing hw13
