+++ hw13/task2/abstract_reader.py
@@ -5,5 +5,3 @@ class Reader(ABC):
     def read_file(self):
         pass
 
-    def write_file(self, text: str):
-        pass
