+++ hw13/task2/abstract_writer.py
@@ -1,0 +1,7 @@
+from abc import ABC, abstractmethod
+
+
+class Writer(ABC):
+    @abstractmethod
+    def write(self, new_data):
+        pass
